Split state/comments/selectors.
Also reimplements some `_.get()` usage with `?.` and `??`.
