+++ client/state/comments/selectors/get-comment-errors.js
@@ -1,0 +1,8 @@
+/**
+ * Internal dependencies
+ */
+import 'state/comments/init';
+
+export function getCommentErrors( state ) {
+	return state.comments.errors;
+}
