+++ client/state/comments/selectors/get-expansions-for-post.js
@@ -1,0 +1,10 @@
+/**
+ * Internal dependencies
+ */
+import { getStateKey } from 'state/comments/utils';
+
+import 'state/comments/init';
+
+export function getExpansionsForPost( state, siteId, postId ) {
+	return state.comments.expansions[ getStateKey( siteId, postId ) ];
+}
